separated safe area element from safe area controller, made safe area element registration opt-in
